+++ core/src/main/java/io/squashql/table/MergeTables.java
@@ -2,20 +2,24 @@ package io.squashql.table;
 
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
-import io.squashql.query.*;
+import io.squashql.query.ColumnarTable;
+import io.squashql.query.Header;
+import io.squashql.query.Measure;
+import io.squashql.query.Table;
 import io.squashql.query.database.SQLTranslator;
-import io.squashql.util.NullAndTotalComparator;
+import io.squashql.query.dictionary.ObjectArrayDictionary;
+import io.squashql.query.dto.JoinType;
+import org.eclipse.collections.api.list.primitive.IntList;
+import org.eclipse.collections.api.list.primitive.MutableIntList;
+import org.eclipse.collections.api.set.primitive.MutableIntSet;
+import org.eclipse.collections.impl.list.mutable.primitive.MutableIntListFactoryImpl;
+import org.eclipse.collections.impl.set.mutable.primitive.IntHashSet;
+import org.eclipse.collections.impl.set.mutable.primitive.MutableIntSetFactoryImpl;
 
 import java.util.*;
+import java.util.stream.IntStream;
 
-public final class MergeTables {
-
-  // FIXME dirty fix until https://github.com/squashql/squashql/issues/127 is done
-  static final Comparator<?> comp = NullAndTotalComparator.nullsLastAndTotalsFirst(Comparator.naturalOrder());
-
-  static <T> int compare(T a , T b) {
-    return ((Comparator<T>) comp).compare(a, b);
-  }
+public class MergeTables {
 
   private MergeTables() {
   }
@@ -29,38 +33,53 @@ public final class MergeTables {
   }
 
   /**
-   * Merge two tables into only one resulting table. Hypothesis: the input tables must not share any measure.
+   * Same as {@link #mergeTables(Table, Table, JoinType)} but with {@code JoinType = FULL}
    */
   public static Table mergeTables(Table leftTable, Table rightTable) {
-    if (leftTable == null) {
+    return mergeTables(leftTable, rightTable, JoinType.FULL);
+  }
+
+  /**
+   * Merge two tables into only one resulting table. We choose to first get all columns and then all measures.
+   * Hypothesis:
+   * <ul>
+   *   <li>the input tables must have their common columns at the beginning, in the same order, and sorted</li>
+   *   <li>every table must have all its columns first, then its measures</li>
+   *   <li>the input tables must not share any measure</li>
+   * </ul>
+   */
+  public static Table mergeTables(Table leftTable, Table rightTable, JoinType joinType) {
+    Objects.requireNonNull(leftTable);
+    Objects.requireNonNull(rightTable);
+
+    if (leftTable.count() == 0) { // empty
       return rightTable;
     }
-    if (rightTable == null) {
+    if (rightTable.count() == 0) { // empty
       return leftTable;
     }
 
-    final List<Header> mergedTableHeaders = mergeHeaders(leftTable, rightTable);
+    final Holder mergedTableHeaders = mergeHeaders(leftTable, rightTable);
     final Set<Measure> mergedTableMeasures = mergeMeasures(leftTable.measures(), rightTable.measures());
-    final List<List<Object>> mergedValues = mergeValues(mergedTableHeaders, leftTable, rightTable);
+    final List<List<Object>> mergedValues = mergeValues(mergedTableHeaders, leftTable, rightTable, joinType);
 
     return new ColumnarTable(
-            mergedTableHeaders,
+            mergedTableHeaders.headers,
             mergedTableMeasures,
             mergedValues);
   }
 
-  private static int getCommonColumnsCount(List<Header> leftHeaders, List<Header> rightHeaders) {
-    return (int) leftHeaders.stream().filter(rightHeaders::contains).count();
-  }
-
-  private static List<Header> mergeHeaders(Table leftTable, Table rightTable) {
+  private static Holder mergeHeaders(Table leftTable, Table rightTable) {
     List<Header> mergedColumns = new ArrayList<>();
     List<Header> mergedMeasures = new ArrayList<>();
+    MutableIntList leftMappingList = MutableIntListFactoryImpl.INSTANCE.empty();
+    MutableIntList rightMappingList = MutableIntListFactoryImpl.INSTANCE.empty();
     leftTable.headers().forEach(leftHeader -> {
       if (leftHeader.isMeasure()) {
         mergedMeasures.add(leftHeader);
       } else {
         mergedColumns.add(leftHeader);
+        leftMappingList.add(mergedColumns.indexOf(leftHeader));
       }
     });
     rightTable.headers().forEach(rightHeader -> {
@@ -76,169 +95,210 @@ public final class MergeTables {
         if (!mergedColumns.contains(rightHeader)) {
           mergedColumns.add(rightHeader);
         }
+        rightMappingList.add(mergedColumns.indexOf(rightHeader));
       }
     });
+
     List<Header> mergedTableHeaders = new ArrayList<>(mergedColumns);
     mergedTableHeaders.addAll(mergedMeasures);
-    return mergedTableHeaders;
+    return new Holder(leftTable, rightTable, mergedTableHeaders, leftMappingList, rightMappingList);
   }
 
   private static Set<Measure> mergeMeasures(Set<Measure> leftMeasures, Set<Measure> rightMeasures) {
     return Sets.newHashSet(Iterables.concat(leftMeasures, rightMeasures));
   }
 
-  /**
-   * In order to merge the tables, they must have their common columns at the beginning, in the same order, and sorted.
-   * They also must have all their columns first, then their measures.
-   *
-   * This method takes as input the tables to merge and return them reshaped to respect this.
-   */
-  private static Table[] prepareTablesForMerge(Table leftTable, Table rightTable) {
-    // Order columns to have all commons first
-    List<Header> leftHeaders = leftTable.headers();
-    List<Header> rightHeaders = rightTable.headers();
-    List<String> commonColumns = leftHeaders.stream()
-            .filter(header -> !header.isMeasure() && rightHeaders.contains(header)).map(header -> header.name())
+  private static List<List<Object>> mergeValues(Holder holder, Table leftTable, Table rightTable, JoinType joinType) {
+    List<Header> mergedTableHeaders = holder.headers;
+    Object[] mergedBuffer = new Object[(int) mergedTableHeaders.stream().filter(h -> !h.isMeasure()).count()];
+//    Object[] leftBuffer = new Object[(int) leftTable.headers().stream().filter(h -> !h.isMeasure()).count()];
+    Object[] rightBuffer = new Object[(int) rightTable.headers().stream().filter(h -> !h.isMeasure()).count()];
+
+    int[] leftMapping = holder.leftMapping;
+    int[] rightMapping = holder.rightMapping;
+
+    long leftCountMeasure = leftTable.headers().stream().filter(Header::isMeasure).count();
+    long rightCountMeasure = rightTable.headers().stream().filter(Header::isMeasure).count();
+    List<ArrayList<Object>> measureValues = IntStream.range(0, (int) (leftCountMeasure + rightCountMeasure))
+            .mapToObj(i -> new ArrayList<>()).toList();
+    List<ArrayList<Object>> pointValues = IntStream.range(0, mergedBuffer.length)
+            .mapToObj(i -> new ArrayList<>()).toList();
+
+    BitSet alreadyVisited = new BitSet();
+
+    List<Measure> leftMeasures = getMeasures((ColumnarTable) leftTable);
+    List<Measure> rightMeasures = getMeasures((ColumnarTable) rightTable);
+    final int[] position = {-1};
+
+    leftTable.pointDictionary().forEach((point, row) -> {
+      Arrays.fill(mergedBuffer, SQLTranslator.TOTAL_CELL);
+      for (int i = 0; i < point.length; i++) {
+        mergedBuffer[leftMapping[i]] = point[i];
+      }
+
+      if (joinType == JoinType.INNER && !holder.doesGlobalPointExistInRightTable(mergedBuffer)) {
+        return; // abort
+      }
+
+      for (int i = 0; i < rightMapping.length; i++) {
+        rightBuffer[i] = mergedBuffer[rightMapping[i]];
+      }
+
+      for (int i = 0; i < mergedBuffer.length; i++) {
+        pointValues.get(i).add(mergedBuffer[i]);
+      }
+
+      for (int i = 0; i < leftCountMeasure; i++) {
+        measureValues.get(i).add(leftTable.getAggregateValues(leftMeasures.get(i)).get(row));
+      }
+
+      int[] complement = complement(holder.leftMapping, holder.rightMapping);
+      boolean missingColumnsAreTotal = true;
+      for (int c : complement) {
+        missingColumnsAreTotal &= SQLTranslator.TOTAL_CELL.equals(mergedBuffer[c]);
+      }
+
+      boolean pointExistInRightTable = (position[0] = rightTable.pointDictionary().getPosition(rightBuffer)) >= 0;
+      boolean condition = (holder.leftIsIncludeInRight || missingColumnsAreTotal) && pointExistInRightTable;
+      for (int i = 0; i < rightCountMeasure; i++) {
+        Object value = null;
+        if (condition) {
+          value = rightTable.getAggregateValues(rightMeasures.get(i)).get(position[0]);
+        }
+        measureValues.get((int) (i + leftCountMeasure)).add(value);
+      }
+
+      if (condition) {
+        alreadyVisited.set(position[0]);
+      }
+    });
+
+    rightTable.pointDictionary().forEach((point, row) -> {
+      if (alreadyVisited.get(row)) {
+        return;
+      }
+      Arrays.fill(mergedBuffer, SQLTranslator.TOTAL_CELL);
+      for (int i = 0; i < point.length; i++) {
+        mergedBuffer[rightMapping[i]] = point[i];
+      }
+
+      if ((joinType == JoinType.LEFT || joinType == JoinType.INNER) && !holder.doesGlobalPointExistInLeftTable(mergedBuffer)) {
+        return; // abort
+      }
+
+      for (int i = 0; i < mergedBuffer.length; i++) {
+        pointValues.get(i).add(mergedBuffer[i]);
+      }
+
+      for (int i = 0; i < leftCountMeasure; i++) {
+        measureValues.get(i).add(null);
+      }
+
+      for (int i = 0; i < rightCountMeasure; i++) {
+        measureValues.get((int) (i + leftCountMeasure)).add(rightTable.getAggregateValues(rightMeasures.get(i)).get(row));
+      }
+    });
+
+    List<List<Object>> result = new ArrayList<>(pointValues);
+    result.addAll(measureValues);
+    return result;
+  }
+
+  private static List<Measure> getMeasures(ColumnarTable table) {
+    return table.headers().stream()
+            .filter(Header::isMeasure)
+            .map(h -> table.measures().stream().filter(m -> m.alias().equals(h.name())).findFirst().orElseThrow(() -> new IllegalStateException("Cannot find measure with name " + h.name())))
             .toList();
-
-    List<String> leftColumns = leftHeaders.stream().filter(header -> !header.isMeasure())
-            .map(header -> header.name()).toList();
-    List<String> leftColumnsOrdered = new ArrayList<>(commonColumns.stream().filter(leftColumns::contains).toList());
-    leftColumns.forEach(columnName -> {
-      if (!leftColumnsOrdered.contains(columnName)) {
-        leftColumnsOrdered.add(columnName);
-      }
-    });
-
-    List<String> rightColumns = rightHeaders.stream().filter(header -> !header.isMeasure())
-            .map(header -> header.name()).toList();
-    List<String> rightColumnsOrdered = new ArrayList<>(commonColumns.stream().filter(rightColumns::contains).toList());
-    rightColumns.forEach(columnName -> {
-      if (!rightColumnsOrdered.contains(columnName)) {
-        rightColumnsOrdered.add(columnName);
-      }
-    });
-
-    ColumnarTable orderedLeftTable = TableUtils.selectAndOrderColumns((ColumnarTable) leftTable, leftColumnsOrdered,
-            leftTable.measures().stream().toList());
-    ColumnarTable orderedRightTable = TableUtils.selectAndOrderColumns((ColumnarTable) rightTable, rightColumnsOrdered,
-            rightTable.measures().stream().toList());
-
-    // Sort rows on all columns with default comparator (natural order)
-    orderedLeftTable = (ColumnarTable) TableUtils.orderRows(orderedLeftTable, Collections.emptyMap(),
-            Collections.emptySet());
-    orderedRightTable = (ColumnarTable) TableUtils.orderRows(orderedRightTable, Collections.emptyMap(),
-            Collections.emptySet());
-
-    return new Table[] {orderedLeftTable, orderedRightTable};
   }
 
-  private static List<List<Object>> mergeValues(
-          List<Header> mergedTableHeaders,
-          Table leftTable,
-          Table rightTable) {
-    // Order headers and sort rows
-    Table[] preparedTablesForMerge = prepareTablesForMerge(leftTable, rightTable);
-    leftTable = preparedTablesForMerge[0];
-    rightTable = preparedTablesForMerge[1];
-    // values initialization
-    final List<List<Object>> mergedValues = new ArrayList<>();
-    for (int i = 0; i < mergedTableHeaders.size(); i++) {
-      mergedValues.add(new ArrayList<>());
+  static class Holder {
+
+    final List<Header> headers;
+    final int[] leftMapping;
+    final int[] rightMapping;
+    final boolean leftIsIncludeInRight;
+    final int[] intersection;
+    final Table leftTable;
+    final Table rightTable;
+    ObjectArrayDictionary leftTableCommonPointDic; // computed on demand
+    ObjectArrayDictionary rightTableCommonPointDic; // computed on demand
+
+    public Holder(Table leftTable, Table rightTable, List<Header> headers, IntList leftMapping, IntList rightMapping) {
+      this.leftTable = leftTable;
+      this.rightTable = rightTable;
+      this.headers = headers;
+      this.leftMapping = leftMapping.toArray();
+      this.rightMapping = rightMapping.toArray();
+      this.leftIsIncludeInRight = isLeftListIncludeInRightList(leftMapping, rightMapping);
+      this.intersection = intersection(this.leftMapping, this.rightMapping);
     }
 
-    List<Header> leftHeaders = leftTable.headers();
-    List<Header> rightHeaders = rightTable.headers();
-    int commonColumnsCount = getCommonColumnsCount(leftHeaders, rightHeaders);
+    private ObjectArrayDictionary buildIntersectionPointDictionary(Table table, int[] mapping) {
+      ObjectArrayDictionary dictionary = new ObjectArrayDictionary(this.intersection.length);
+      int[] interMapping = new int[this.intersection.length];
+      for (int i = 0; i < this.intersection.length; i++) {
+        interMapping[i] = indexOf(mapping, this.intersection[i]);
+      }
 
-    int leftRowIndex = 0;
-    int rightRowIndex = 0;
-    List<Object> leftRow = leftTable.getFactRow(leftRowIndex);
-    List<Object> rightRow = rightTable.getFactRow(rightRowIndex);
-    while (leftRow != null || rightRow != null) {
-      MergeRowsStrategy mergeRowsStrategy = getMergeRowsStrategy(leftRow, rightRow, commonColumnsCount);
-      switch (mergeRowsStrategy) {
-        case KEEP_LEFT -> {
-          addRowFromTableToValues(mergedValues, mergedTableHeaders, leftTable, leftRowIndex);
-          leftRow = leftTable.getFactRow(++leftRowIndex);
+      table.forEach(row -> {
+        Object[] columnValues = new Object[this.intersection.length];
+        for (int i = 0; i < this.intersection.length; i++) {
+          columnValues[i] = row.get(interMapping[i]);
         }
-        case KEEP_RIGHT -> {
-          addRowFromTableToValues(mergedValues, mergedTableHeaders, rightTable, rightRowIndex);
-          rightRow = rightTable.getFactRow(++rightRowIndex);
-        }
-        case MERGE -> {
-          addMergedRowToValues(mergedValues, mergedTableHeaders, leftTable, leftRowIndex, rightTable, rightRowIndex);
-          leftRow = leftTable.getFactRow(++leftRowIndex);
-          rightRow = rightTable.getFactRow(++rightRowIndex);
-        }
+        dictionary.map(columnValues);
+      });
+      return dictionary;
+    }
+
+    public boolean doesGlobalPointExistInLeftTable(Object[] globalPoint) {
+      if (this.leftTableCommonPointDic == null) {
+        this.leftTableCommonPointDic = buildIntersectionPointDictionary(this.leftTable, this.leftMapping);
       }
+      return doesGlobalPointExist(globalPoint, this.leftTableCommonPointDic);
     }
 
-    return mergedValues;
-  }
-
-  enum MergeRowsStrategy {
-    KEEP_LEFT,
-    KEEP_RIGHT,
-    MERGE
-  }
-
-  private static MergeRowsStrategy getMergeRowsStrategy(List<Object> leftRow, List<Object> rightRow,
-          int commonColumnsCount) {
-    // Handle null row cases
-    if (leftRow == null) {
-      return MergeRowsStrategy.KEEP_RIGHT;
-    }
-    if (rightRow == null) {
-      return MergeRowsStrategy.KEEP_LEFT;
-    }
-
-    // Check if the two rows have the same values on the common columns
-    for (int commonIndex = 0; commonIndex < commonColumnsCount; commonIndex++) {
-      Object leftRowValue = leftRow.get(commonIndex);
-      Object rightRowValue = rightRow.get(commonIndex);
-      if (!leftRowValue.equals(rightRowValue)) { // The two rows don't have the same values on all the common columns
-        // Keep one row ensuring that the merged table will be well sorted
-        // Is it enough to compare string values ?
-        return compare(leftRowValue, rightRowValue) < 0 ? MergeRowsStrategy.KEEP_LEFT : MergeRowsStrategy.KEEP_RIGHT;
+    public boolean doesGlobalPointExistInRightTable(Object[] globalPoint) {
+      if (this.rightTableCommonPointDic == null) {
+        this.rightTableCommonPointDic = buildIntersectionPointDictionary(this.rightTable, this.rightMapping);
       }
+      return doesGlobalPointExist(globalPoint, this.rightTableCommonPointDic);
     }
-    // The two rows have the same values on all the common columns
-    // Check if there are only total values on the non-common columns
-    if (((leftRow.size() == commonColumnsCount) || leftRow.subList(commonColumnsCount, leftRow.size()).stream()
-            .noneMatch(x -> x != SQLTranslator.TOTAL_CELL))
-            && ((rightRow.size() == commonColumnsCount) || rightRow.subList(commonColumnsCount, rightRow.size())
-            .stream().noneMatch(x -> x != SQLTranslator.TOTAL_CELL))) {
-      // We can merge the rows
-      return MergeRowsStrategy.MERGE;
-    }
-    // Keep one row ensuring that the merged table will be well sorted
-    return leftRow.size() == commonColumnsCount ? MergeRowsStrategy.KEEP_LEFT : MergeRowsStrategy.KEEP_RIGHT;
-  }
 
-  private static void addRowFromTableToValues(List<List<Object>> values, List<Header> headers, Table table,
-          int rowToAddIndex) {
-    for (int index = 0; index < headers.size(); index++) {
-      Header header = headers.get(index);
-      Object element = header.isMeasure() ? null : SQLTranslator.TOTAL_CELL;
-      if (table.headers().contains(header)) {
-        element = table.getColumnValues(header.name()).get(rowToAddIndex);
+    private boolean doesGlobalPointExist(Object[] globalPoint, ObjectArrayDictionary dictionary) {
+      Object[] array = new Object[this.intersection.length];
+      for (int i = 0; i < this.intersection.length; i++) {
+        array[i] = globalPoint[this.intersection[i]];
       }
-      values.get(index).add(element);
+      return dictionary.getPosition(array) >= 0;
     }
   }
 
-  private static void addMergedRowToValues(List<List<Object>> values, List<Header> headers, Table leftTable,
-          int leftRowIndex, Table rightTable, int rightRowIndex) {
-    for (int index = 0; index < headers.size(); index++) {
-      Object element;
-      Header header = headers.get(index);
-      if (leftTable.headers().contains(header)) {
-        element = leftTable.getColumnValues(header.name()).get(leftRowIndex);
-      } else {
-        element = rightTable.getColumnValues(header.name()).get(rightRowIndex);
+  private static boolean isLeftListIncludeInRightList(IntList left, IntList right) {
+    MutableIntSet s = new IntHashSet(left);
+    s.removeAll(right);
+    return s.isEmpty();
+  }
+
+  public static int[] complement(int[] a, int[] b) {
+    MutableIntSet s1 = MutableIntSetFactoryImpl.INSTANCE.of(a);
+    MutableIntSet s2 = MutableIntSetFactoryImpl.INSTANCE.of(b);
+    s1.removeAll(s2);
+    return s1.toArray();
+  }
+
+  public static int[] intersection(int[] a, int[] b) {
+    MutableIntSet s1 = MutableIntSetFactoryImpl.INSTANCE.of(a);
+    MutableIntSet s2 = MutableIntSetFactoryImpl.INSTANCE.of(b);
+    s1.retainAll(s2);
+    return s1.toArray();
+  }
+
+  private static int indexOf(int[] a, int i) {
+    for (int j = 0; j < a.length; j++) {
+      if (a[j] == i) {
+        return j;
       }
-      values.get(index).add(element);
     }
+    throw new IllegalArgumentException();
   }
 }
